Make blocks arguments to interop methods, automatically promote Procs to delegates
If a block is given to an interop method the proc is pushed
as the last argument. So the following become equivalent:

  obj.meth { ... }
  obj.meth(Proc { ... })

When an argument containing a Proc is matched with a parameter
that has a delegate type, the Proc is automatically promoted
to a matching delegate instance.

In winforms we can then write:

  button.add_Click { |sender,args| do_something }


git-svn-id: https://rubydotnetcompiler.googlecode.com/svn/trunk@56 f224b4e6-8335-0410-baf7-2d8196006633